Move and update references to list_pdf_utils
