+++ common/__init__.py
@@ -1,4 +1,2 @@
-#from .utils import extract_info
-from .utils import get_files
 from .base_scrapers.opendata import opendata_scraper
 from .base_scrapers.opendata import opendata_scraper2
