+++ common/utils/__init__.py
@@ -1,4 +1,2 @@
 from .file_downloaders.downloaders import *
-from .list_pdf_utils import get_files
-#from .list_pdf_utils import extract_info
 from .website_hasher.page_update import hash_comparer, page_hasher, page_update
