--- common/utils/list_pdf_utils/__init__.py
@@ -1,2 +1,0 @@
-#from .extract_info import extract_info
-from .get_files import get_files
