+++ examples_templates/scraper_template/scraper.py
@@ -17,7 +17,7 @@ from pathlib import Path
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
 
-from common.utils import get_files # example of importing one of those files
+from utils.pdf.list_pdf_utils.get_files import get_files # example of importing one of those files
 
 
 # Put your code here
