+++ scrapers/CA/contra_costa_county/municipality/antioch/antioch_scraper.py
@@ -9,7 +9,7 @@ from pathlib import Path
 
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
-from common import get_files
+from utils.pdf.list_pdf_utils.get_files import get_files
 
 webpage = "https://www.antiochca.gov/police/crime-statistics/crime-statistics/"
 """
