+++ scrapers/CA/solano_county/municipal/dixon/response_extractor.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
-from common.list_pdf_utils import get_files
+from utils.pdf.list_pdf_utils.get_files import get_files
 
 save_dir = "./data/"
