+++ scrapers/GA/whitfield_county/municipal/Dalton/dalton_scraper.py
@@ -9,7 +9,7 @@ import sys
 
 p = Path(__file__).resolve().parents[5]
 sys.path.insert(1, str(p))
-from common.utils import get_files
+from utils.pdf.list_pdf_utils.get_files import get_files
 
 
 cur_dir = os.getcwd()
